Updated Order.java
Changed quite a few things, adding and removing items from the cart are now done incrementally, and use the order id rather than the order object. Rather than using the store name, we now use the storeId. Added getters and setters for all of the variables in the class. Added a boolean flag as a marker for whether or not the order is completed. Added a sendOrder method that writes the current order to the firebase database.